Modify: JPA 엔티티 연관관계 정리 (#5)
Review 테이블 상속 테이블 조인 전략으로 분리

+++ src/main/java/com/livable/server/entity/BuildingRestaurantMap.java
@@ -32,5 +32,5 @@ public class BuildingRestaurantMap extends BaseTimeEntity {
     private Restaurant restaurant;
 
     @Column(nullable = false)
-    private Long distance;
+    private Integer distance;
 }
